Use custom folder for picking images
Search the first class name (i.e. the first with a 'Page' suffix) in the
object hierarchy that has a correspoding folder in ASSETS_PATH, that is
a folder with the same name but without the 'Page' suffix. This folder
will be used as custom path.

For example, if this class is `HomePage` and it is inherited from
`CarouselPage`, will be checked `assets/Home` first and
`assets/Carousel` after.

If no valid folders are found, the default folder is used.