Downgrade sortablefile dependency
Previously silverstripe-gallery was depending on a specific commit of
bummzack/sortablefile. This was generating a lot of fuss on composer
that refuses to install (I don't know why) this dependency:

$ composer require entidi/silverstripe-carousel
Using version ^2.2 for entidi/silverstripe-carousel
./composer.json has been updated
Loading composer repositories with package information
Updating dependencies (including require-dev)
Your requirements could not be resolved to an installable set of packages.

  Problem 1
    - entidi/silverstripe-carousel 2.2.1 requires bummzack/sortablefile dev-master#48c3fb4c6b37db4e109ec5558e3db6d20c82a75b -> no matching package found.
    - entidi/silverstripe-carousel 2.2.0 requires bummzack/sortablefile dev-master#48c3fb4c6b37db4e109ec5558e3db6d20c82a75b -> no matching package found.
    - Installation request for entidi/silverstripe-carousel ^2.2 -> satisfiable by entidi/silverstripe-carousel[2.2.0, 2.2.1].

[... other composer crap cut out ...]